better libgxbuild

+++ include/FlashImage.hpp
@@ -44,7 +44,7 @@ typedef struct _raw_nand_header {
 
 typedef struct BlResults {
     uint32_t offset;
-    uint8_t version;
+    uint16_t version;
     uint32_t size;
     uint32_t magic;
 } bl_results_t;
